add offset time function

+++ inc/alljoyn/audio/SinkPlayer.h
@@ -211,7 +211,7 @@ class SinkPlayer : public ajn::MessageReceiver {
     size_t GetSinkCount();
 
 
-    bool addoffset(const char* name, int offset);
+    bool addoffset(const char* name, int64_t offset);
     /**
      * Opens the stream to the sink.
      *
